HSEARCH-2003 Considering field name when configuring numeric fields through the DSL

+++ orm/src/test/java/org/hibernate/search/test/configuration/Item.java
@@ -18,7 +18,7 @@ public class Item {
 	@Id @GeneratedValue
 	private Integer id;
 	private String description;
-	private Double price;
+	private Short price;
 
 
 	@ManyToOne( cascade = { CascadeType.PERSIST, CascadeType.REMOVE } )
@@ -49,11 +49,11 @@ public class Item {
 		this.description = description;
 	}
 
-	public Double getPrice() {
+	public Short getPrice() {
 		return price;
 	}
 
-	public void setPrice(Double price) {
+	public void setPrice(Short price) {
 		this.price = price;
 	}
 }
